// src/interfaces/group.ts
import { User } from './user';
// Generated by https://quicktype.io


export interface GroupResponse {
  groups?: Group[];
  group?: Group;
}

export interface Group {
  id:          string;
  name:        string;
  description: string;
  author:      User;
  users:       User[];
  usersKill:   UsersKill[];
  groupKills:  number;
}


export interface UsersKill {
  user:  User;
  kills: number;
}